hopefully fix all remaining tests
